Hide deprecation warning with createVault constructor
This is a cleanup patch from the new constructor feature.

The `KeyStore.createVault` method relies on the underlying `new` constructor, which we have issuing a deprecation warning, so now that warning is actually firing always.

We were talking about getting rid of the `new` method eventually, but to actually stop using it while most of its methods are instance methods on the prototype would require a bit of a refactor.

This is a cheap solution that avoids that refactor for the moment.

In the slightly longer term, we might want to make the current KeyStore class simply return another internal class, so it can use the `new` constructor internally, while only throwing the deprecation warning when its own new constructor is used.

I'm opening to making that change to this PR, but wanted to open the discussion first.